update embedded view samples to use nuget libs
