--- Basic Samples/MultiViewSample/EmbeddedSample/AndroidManifest.xml
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.EmbeddedSample" android:versionCode="1" android:versionName="1.0" android:installLocation="auto">
-	<uses-sdk android:minSdkVersion="21" android:targetSdkVersion="30" />
-	<application android:label="EmbeddedSample"></application>
-</manifest>
