+++ Basic Samples/MultiViewSample/MultiViewSample/.config/dotnet-tools.json
