+++ Basic Samples/MultiViewSample/MultiViewSample/Content/Content.mgcb
