+++ Basic Samples/MultiViewSample/MultiViewSample/Resources/Drawable/Icon.png
