+++ Basic Samples/MultiViewSample/MultiViewSample/Resources/Drawable/Splash.png
