+++ Basic Samples/MultiViewSample/MultiViewSample/Resources/Layout/Main.xml
