+++ Basic Samples/MultiViewSample/MultiViewSample/Resources/Values/Strings.xml
