+++ Basic Samples/MultiViewSample/MultiViewSample/Resources/Values/Styles.xml
